+++ application/routes/users.routes.ts
@@ -1,0 +1,14 @@
+import { Router } from "express";
+
+const router = Router();
+
+router.get('/', async (req, res, next) => {
+    try {
+      res.json('exelent');
+    } catch (e) {
+        res.sendStatus(400);
+      next(e);
+    }
+});
+
+export default router;